+++ docs/docs/CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed CEL expression matching validator to now properly error out when it receives empty expressions
 - Added OpenRC init.d script.
 - Added `--version` flag.
+- Added `anubis_proxied_requests_total` metric to count proxied requests.
 
 ## v1.18.0: Varis zos Galvus